필요없는 코드 제거, Code Formatting, DTO 형태 만들어 수정, 폴더 구조 변경
