+++ docs/requirements.md
@@ -32,17 +32,21 @@ public enum FormItemType {
 
 설문조사 생성
 - 설문조사 생성 시 설문 받을 항목의 형태에 따른 밸리데이션(형태에 따른 하위 항목에 대한 유효성, 선택리스트의 경우 해당 리스트가 포함 되었는지?), 기타 칼럼 길이에 대한 밸리데이션 정도 필요할 것.
+- 하위 항목에 대한 유효성 검사
+  - 롱텍스트(500자 제한), 텍스트(500자 제한(?)), 단일,다중 선택지의 경우 선택지의 유무 체크 정도 진행.
 
 설문조사 수정 
 - 설문조사 이름, 설명, 설문 받을 항목 
-- 설문조사의 이름, 설명 등에 대한 버저닝은 미 진행 예정 
+- 설문조사의 이름, 설명 등에 대한 버저닝은 미 진행 예정, 설문조사 이름, 설명등은 과거 기록이 사라짐.
 - 생성시와 같은 조건의 밸리데이션 
 - 수정 API 발생할 때 마다, 수정된 모든 항목을 검사하여 변경 여부를 체크하기가 어려울 것으로 생각.  
 	- 먼저 업데이트 시 매번 항목에 대한 버저닝을 업데이트 한 항목을 추가로 생성.
-	- 상기 로직은 단 한 개의 항목의 하위 항목이 변경 되어도 모든 항목에 대한 내용이 추가 복제되어 신규 버전으로 추가되는 것으로, 추후 결과 값을 통계 내거나 데이터를 추가 할 시 데이터의 수의 추가 
+	- 상기 로직은 단 한 개의 항목의 하위 항목이 변경 되어도 모든 항목에 대한 내용이 추가 복제되어 신규 버전으로 추가되는 것으로, 추후 결과 값을 통계 내거나 데이터를 추가 할 시 데이터의 수의 추가
 
 설문조사 응답 제출 
 - 항목 밸리데이션
+- 응답 중 수정중인 사람은, 과거 버전과 현재 버전의 응답이 혼재. 
+- But 수정되었다고 기존 것이 안들어 가면 다 작성 한 설문값을 날리게 됨.
 
 설문조사 응답 조회 
 - 설문조사 식별자 필요(설문 ID? 정도로 생각)
