+++ docs/retrospect.md
@@ -1,0 +1,17 @@
+## TODO
+DTO Validation 라이브러리 찾아보기 
+
+DTO 내에서 모델을 가져다 쓰는 부분은 제거, 코드 재사용 성을 위해 DTO? 를 대신할 수 있는 방법을 찾아보자.
+
+이대로 가면 DTO가 수백개가 되겠다!!
+
+자바의 폴더 구조 Common Case 조사해보자
+
+JPA학습. 
+Optional Keyword 확인 
+
+Response 형태를 data, status 형태로만 단순히 일치할 수 있게 수정해보자.
+
+기능 구현 후 개선 해보자!
+
+식별 관계가 JPA 랑 좀 안맞나?
