--- src/main/java/com/chanki/form/domain/forms/Form.java
@@ -1,49 +1,0 @@
-package com.chanki.form.domain.forms;
-
-import java.util.List;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@ToString
-@Getter
-@NoArgsConstructor
-@Entity
-@Table(name = "form")
-public class Form {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "form_id")
-	private Long formID;
-		
-	@Column(length = 100, nullable = false)
-	private String title;
-	
-	@Column(length = 500, nullable = false)
-	private String description;
-	
-	@OneToMany(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-	@JoinColumn(name="form_id", referencedColumnName = "form_id")
-	private List<FormItem> formItems;
-	
-	@Builder
-	public Form(String title, String description, List<FormItem> formItems) {
-		this.title = title;
-		this.description = description;
-		this.formItems = formItems;
-	}
-}
