--- src/main/java/com/chanki/form/domain/forms/FormItemRepository.java
@@ -1,7 +1,0 @@
-package com.chanki.form.domain.forms;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FormItemRepository extends JpaRepository<FormItem, FormItemId> {
-
-}
