+++ src/main/java/com/chanki/form/web/domain/forms/FormItemId.java
@@ -1,4 +1,4 @@
-package com.chanki.form.domain.forms;
+package com.chanki.form.web.domain.forms;
 
 import java.io.Serializable;
 
@@ -6,20 +6,23 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Id;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
+@ToString
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 @NoArgsConstructor
 public class FormItemId implements Serializable {
-	@EqualsAndHashCode.Include
-	@Id
-	@Column(name = "form_id")
-	private long formId;
-	
-	@EqualsAndHashCode.Include
-	@Id
-	private long version;
-	
-	@EqualsAndHashCode.Include
-	@Id
-	private long sequence;
+
+  @EqualsAndHashCode.Include
+  @Id
+  @Column(name = "form_id")
+  private long formId;
+
+  @EqualsAndHashCode.Include
+  @Id
+  private long version;
+
+  @EqualsAndHashCode.Include
+  @Id
+  private long sequence;
 }
