+++ src/main/java/com/chanki/form/web/domain/forms/FormItemOptionRepository.java
@@ -1,4 +1,4 @@
-package com.chanki.form.domain.forms;
+package com.chanki.form.web.domain.forms;
 
 import org.springframework.data.jpa.repository.JpaRepository;
