+++ src/main/java/com/chanki/form/web/domain/forms/FormItemType.java
@@ -1,6 +1,6 @@
-package com.chanki.form.domain.forms;
+package com.chanki.form.web.domain.forms;
 
 // 단답형, 장문형, 단일선택리스트, 다중선택리스트
 public enum FormItemType {
-	TEXT, LONGTEXT, RADIO, CHECKBOX
+  TEXT, LONGTEXT, RADIO, CHECKBOX
 }
