+++ src/main/java/com/chanki/form/web/dto/FormItemOptionCreateRequestDto.java
@@ -1,0 +1,12 @@
+package com.chanki.form.web.dto;
+
+import jakarta.persistence.Column;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class FormItemOptionCreateRequestDto {
+
+  private String description;
+}
